Added LogicAgent.go
